pyproject.toml
[project]
name = "plc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "keras>=3.8.0",
    "matplotlib>=3.10.0",
    "numpy==1.26.4",
    "pandas>=2.2.3",
    "pymodbus>=3.8.0",
    "pymso5000-tspspi>=0.0.10",
    "pyqt6>=6.8.0",
    "scikit-learn>=1.6.1",
    "tensorflow>=2.18.0",
]
